restore behaviour & tests for v2 queries;v3 queries ready to improve;
